Remove unused alerts

+++ src/main/java/org/libtorrent4j/Vectors.java
@@ -28,7 +28,7 @@ public final class Vectors {
     }
 
     public static byte[] byte_vector2bytes(byte_vector v) {
-        int size = v.size();
+        int size = (int)v.size();
         byte[] arr = new byte[size];
 
         for (int i = 0; i < size; i++) {
@@ -39,7 +39,7 @@ public final class Vectors {
     }
 
     public static void byte_vector2bytes(byte_vector v, byte[] arr) {
-        int size = v.size();
+        int size = (int)v.size();
 
         for (int i = 0; i < size; i++) {
             arr[i] = v.get(i);
@@ -63,7 +63,7 @@ public final class Vectors {
     }
 
     public static int[] int_vector2ints(int_vector v) {
-        int size = v.size();
+        int size = (int)v.size();
         int[] arr = new int[size];
 
         for (int i = 0; i < size; i++) {
@@ -74,7 +74,7 @@ public final class Vectors {
     }
 
     public static long[] int64_vector2longs(int64_vector v) {
-        int size = v.size();
+        int size = (int)v.size();
         long[] arr = new long[size];
 
         for (int i = 0; i < size; i++) {
@@ -85,7 +85,7 @@ public final class Vectors {
     }
 
     public static List<String> string_vector2list(string_vector v) {
-        int size = v.size();
+        int size = (int)v.size();
         ArrayList<String> l = new ArrayList<>(size);
 
         for (int i = 0; i < size; i++) {
